Refactor NullToVisibilityConverter and qualify main window namespaces
